+++ src/components/ui/toaster.tsx
@@ -8,7 +8,7 @@ import {
   ToastTitle,
   ToastViewport,
 } from '@/components/ui/toast'
-import { useToast } from '@/utils/api/hooks/use-toast'
+import { useToast } from '@/hooks/use-toast'
 
 export function Toaster() {
   const { toasts } = useToast()
