+++ src/utils/api/index.ts
@@ -1,2 +1,2 @@
-export * from './hooks/hooks'
-export * from './metacom'
+export * from '@/hooks/hooks'
+export * from '@/utils/api/metacom'
